[REFACTOR]: Controller에 있던 Service로직 서비스레이어로 옮김
