+++ pickme/src/main/java/project/pickme/item/controller/ItemController.java
@@ -1,8 +1,4 @@
-package project.pickme.item;
-
-import java.time.LocalDateTime;
-
-import java.util.List;
+package project.pickme.item.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -12,14 +8,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import lombok.RequiredArgsConstructor;
-import project.pickme.item.constant.Status;
 import project.pickme.item.domain.Category;
-import project.pickme.item.dto.ItemDto;
 import project.pickme.item.dto.ItemFormDto;
-import project.pickme.item.dto.LawDto;
 import project.pickme.item.service.ItemService;
-import project.pickme.item.service.LawService;
-import project.pickme.item.service.S3Service;
 import project.pickme.user.constant.Type;
 
 @Controller
@@ -27,31 +18,22 @@ import project.pickme.user.constant.Type;
 @RequiredArgsConstructor
 public class ItemController {
 
-	private final LawService lawService;
 
 	private final ItemService itemService;
 
-	private final S3Service s3Service;
 
 	@GetMapping("/create")
 	public String getItem(Model model) {
-		List<LawDto> laws = lawService.getLaws();
 		model.addAttribute("item", new ItemFormDto());
-		model.addAttribute("laws", laws);
+		model.addAttribute("laws", itemService.findAllLaws());
 		model.addAttribute("type", Type.values());
 		model.addAttribute("categories", Category.values());
 		return "create-item";
 	}
 
 	@PostMapping("/create")
-	public String insertItem(@ModelAttribute ItemFormDto itemFormDTO) {
-		LocalDateTime start = LocalDateTime.of(itemFormDTO.getStartDate(), itemFormDTO.getStartTime());
-		LocalDateTime end = LocalDateTime.of(itemFormDTO.getEndDate(), itemFormDTO.getEndTime());
-		ItemDto itemDTO = new ItemDto(itemFormDTO.getName(), itemFormDTO.getCode(), itemFormDTO.getTarget(),
-			itemFormDTO.getPrice(), start, end,
-			Status.NOT_OPEN, "gunsan");
-
-		itemService.save(itemDTO);
+	public String insertItem(@ModelAttribute ItemFormDto itemFormDto) {
+		itemService.save(itemFormDto);
 		return "redirect:/customs/item/create";
 	}
